requirements.txt
requests==2.31.0
PyGithub==2.1.1
gitpython==3.1.40
radon==6.0.1
lizard==1.17.10
pathspec==0.11.2